jaz_api_v03/requirements.txt
fastapi == 0.111.0
uvicorn == 0.23.2
gunicorn == 21.2.0
pydantic-settings == 2.0.3
asyncpg == 0.28.0
psycopg2-binary == 2.9.7
SQLAlchemy == 2.0.20
alembic == 1.12.0
email_validator == 2.0.0.post2
python-multipart == 0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
oracledb==1.4.1
pycryptodomex==3.20.0
aiosmtplib==3.0.1
pytz==2024.1